[sql] Split `Dialect`s into their own files
